refactor: AddFriend移動
